+++ pkg/service/whatsapp.go
@@ -1,9 +1,11 @@
 package service
 
+import "github.com/brunoOchoa/whatsapp-lib/pkg/model"
+
 type WhatsAppService interface {
 	SendTextMessage(to []string, message string) error
 	SendTemplateMessage(to []string, templateName string, language string) error
-	GetMessageStatus(messageID string) (string, error)
+	GetMessageStatus(webhookJSON []byte) ([]model.StatusInfo, error)
 }
 
 type whatsApp_service struct {
