+++ pkg/whatsapp/status.go
@@ -1,23 +1,15 @@
 package whatsapp
 
 import (
+	"encoding/json"
 	"fmt"
 
 	"github.com/brunoOchoa/whatsapp-lib/pkg/model"
 )
 
-// StatusInfo representa as informações extraídas do webhook
-type StatusInfo struct {
-	MessageID   string
-	Status      string
-	Timestamp   string
-	RecipientID string
-	Errors      []map[string]interface{}
-}
-
-// ParseStatusFromWebhook extrai status do objeto recebido no webhook
-func ParseStatusFromWebhook(webhook *model.Webhook) ([]StatusInfo, error) {
-	var infos []StatusInfo
+// ParseStatusFromWebhook extrai informações de status de mensagens do webhook do WhatsApp.
+func ParseStatusFromWebhook(webhook *model.Webhook) ([]model.StatusInfo, error) {
+	var infos []model.StatusInfo
 
 	if webhook == nil || len(webhook.Entry) == 0 {
 		return nil, fmt.Errorf("webhook vazio ou inválido")
@@ -26,7 +18,7 @@ func ParseStatusFromWebhook(webhook *model.Webhook) ([]StatusInfo, error) {
 	for _, entry := range webhook.Entry {
 		for _, change := range entry.Changes {
 			for _, status := range change.Value.Statuses {
-				info := StatusInfo{
+				info := model.StatusInfo{
 					MessageID:   status.ID,
 					Status:      status.Status,
 					Timestamp:   status.Timestamp,
@@ -48,15 +40,16 @@ func ParseStatusFromWebhook(webhook *model.Webhook) ([]StatusInfo, error) {
 	return infos, nil
 }
 
-func (c *Client) GetMessageStatus(messageID string) (string, error) {
-	endpoint := fmt.Sprintf("%s/%s/messages/%s", API_BASE_URL, c.ApiVersion, messageID)
-	var result map[string]interface{}
-	err := c.sendRequest("GET", endpoint, nil, &result)
-	if err != nil {
-		return "", err
+func (c *Client) GetMessageStatus(webhookJSON []byte) ([]model.StatusInfo, error) {
+	var webhookObj model.Webhook
+	if err := json.Unmarshal(webhookJSON, &webhookObj); err != nil {
+		return nil, fmt.Errorf("erro ao decodificar webhook: %w", err)
 	}
 
-	// Ajuste conforme o formato real da resposta da API
-	status, _ := result["status"].(string)
-	return status, nil
+	infos, err := ParseStatusFromWebhook(&webhookObj)
+	if err != nil {
+		return nil, err
+	}
+
+	return infos, nil
 }
